requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-extensions==3.2.0
django-filter==22.1
djangorestframework==3.13.1
drf-yasg==1.21.3
Faker==14.2.0
flake8==4.0.1
idna==3.3
importlib-metadata==4.2.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
packaging==21.3
pkg-resources==0.0.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
validate-docbr==1.10.0
zipp==3.8.1
